Db implimentation
In the latest commit, i added all api's and my database schema. The DB is fully functional and connects to mongodb over google servers. I am (for now) populating it using user entries, but will change this to be populated with alternative items when I make a fully functional application.